requirements.txt
absl-py==1.3.0
albumentations==1.3.0
backcall==0.2.0
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
einops==0.6.0
ffmpeg==1.4
fonttools==4.37.4
FrEIA==0.2
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
imageio==2.22.2
imgaug==0.4.0
importlib-metadata==5.0.0
iopath==0.1.10
ipython==7.34.0
jedi==0.18.1
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.1.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.6.0
prefetch-generator==1.0.1
prompt-toolkit==3.0.32
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-ignite==0.4.10
pytz==2022.4
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
Shapely==1.8.5.post1
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.3.2
torch @ file:///C:/Users/Administrator/Downloads/torch-1.7.0%2Bcu110-cp37-cp37m-win_amd64.whl
torchsummary==1.5.1
torchvision @ file:///C:/Users/Administrator/Downloads/torchvision-0.8.1%2Bcu110-cp37-cp37m-win_amd64.whl
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
win32-setctime==1.1.0
wincertstore==0.2
yacs==0.1.8
zipp==3.10.0
